Remove release channel from API for now
While the release channels feature is sketched out at a basic level and
even accounted for in the directory server's database design, we aren't
yet certain how it will work, so it would be prudent to not commit to a
specific design in the API before the feature is ready. Besides, there
is no reason to commit to the release channel API before it is fully
implemented, so it's safer to wait to implement it in the API when it's
needed.