Add package info to `AppDownloadInfo`
There are many cases in which a client may want to know the exact
package information for the app described in an AppDownloadInfo. While
clients can access package info for the app in a way by calling
GetAppPackageInfo() with the same app ID passed to GetAppDownloadInfo(),
there is no guarantee that the app won't be updated between these calls,
meaning a client attempting to correlate package info with download info
cannot guarantee that they match.

Therefore, include package info in AppDownloadInfo so that clients can
atomically retrieve both package info and download info in the same
request to guarantee they correlate to each other.