Change compatibility level from an enum to a bool
We are using the compatibility level to primarily determine one thing:
whether the client should attempt to install an app or not. This
decision can be impacted by numerous factors, including but not limited
to:

- whether the app can install successfully based on e.g. min SDK
- whether the app will execute properly based on e.g. supported CPU
  architectures
- whether the app has all functionality limited based on e.g. Play
  Integrity checks

For stability reasons, we can never introduce a new compatibility level
beyond incompatible and compatible. And because there is no real
compatibility "level" beyond these two states (although it can be
impacted by numerous specific causes), it doesn't make sense to
represent it as an enum.

Therefore, change the compatibility level to a simple boolean to better
reflect the data it represents.

We should still hold compatibility information in a message for future
extension possibilities such as fields which indicate whether the
compatibility state is overridable or the reason for an incompatibility.