Revert "Add package info to `AppDownloadInfo`"
This reverts commit 85551c04101d5b75a13ee35f29727cf94fa870bd.

Clients may want to know the exact package information for the app
described in an AppDownloadInfo, as described in the referenced commit.
However, the Accrescent client has no need of this field at the moment,
and it may actually be better to include package info regardless of
whether the app is compatible (contrary to the referenced commit) if we
were to include it.

Since we don't need this information, exclude it from the public API so
that we don't prematurely commit to providing it.